+++ app/src/main/java/com/codesherpa/beerdispenser/app/controllers/BeerController.java
@@ -5,6 +5,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -25,10 +26,12 @@ import jakarta.validation.constraints.Min;
 import com.codesherpa.beerdispenser.app.exceptions.ExceptionMessage;
 import com.codesherpa.beerdispenser.app.exceptions.ServerException;
 
+import java.util.LinkedList;
 import java.util.List;
 
 @RestController
 @RequestMapping("/beers")
+@Validated
 public class BeerController {
     Logger logger = LoggerFactory.getLogger(BeerController.class);
 
@@ -37,35 +40,40 @@ public class BeerController {
 
     @GetMapping
     public ResponseEntity<Object> getAllBeers() {
+        List<ServerException> exceptions = new LinkedList<>();
         try {
             List<Beer> beers = beerService.getAllBeers();
             return new ResponseEntity<>(beers.stream().map(ApiHelper::toBeerDto).toList(), HttpStatus.OK);
         } catch (Exception e) {
             logger.error(ExceptionMessage.BEER_LIST_500, e);
-            return new ResponseEntity<>(new ServerException(ExceptionMessage.BEER_LIST_500), HttpStatus.INTERNAL_SERVER_ERROR);
+            exceptions.add(new ServerException(ExceptionMessage.BEER_LIST_500));
+            return new ResponseEntity<>(exceptions, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<Object> getBeerById(
         @PathVariable @Min(value = 1, message = ExceptionMessage.BEER_ID_INVALID)Long id) {
+        List<ServerException> exceptions = new LinkedList<>();
         try {
             Beer beer = beerService.getBeer(id);
             if (beer != null) {
                 BeerDto beerDto = ApiHelper.toBeerDto(beer);
                 return new ResponseEntity<>(beerDto, HttpStatus.OK);
             } else {
-                return new ResponseEntity<>("", HttpStatus.NOT_FOUND);
+                exceptions.add(new ServerException(ExceptionMessage.BEER_NOT_FOUND));
+                return new ResponseEntity<>(exceptions, HttpStatus.NOT_FOUND);
             }
         } catch (Exception e) {
             logger.error(ExceptionMessage.BEER_GET_500, e);
-            return new ResponseEntity<>(new ServerException(ExceptionMessage.BEER_GET_500 + id),
-                    HttpStatus.INTERNAL_SERVER_ERROR);
+            exceptions.add(new ServerException(ExceptionMessage.BEER_GET_500 + id));
+            return new ResponseEntity<>(exceptions, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @PostMapping
     public ResponseEntity<Object> addBeer(@Valid @RequestBody CreateBeerDto beerDto) {
+        List<ServerException> exceptions = new LinkedList<>();
         Beer beer = new Beer();
         beer.setName(beerDto.name);
         beer.setPricePerLitre(beerDto.pricePerLitre);
@@ -74,20 +82,27 @@ public class BeerController {
             return new ResponseEntity<>(ApiHelper.toBeerDto(newBeer), HttpStatus.CREATED);
         } catch (Exception e) {
             logger.error(ExceptionMessage.BEER_POST_500, e);
-            return new ResponseEntity<>(new ServerException(ExceptionMessage.BEER_POST_500), HttpStatus.INTERNAL_SERVER_ERROR);
+            exceptions.add(new ServerException(ExceptionMessage.BEER_POST_500));
+            return new ResponseEntity<>(exceptions, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Object> deleteBeer(
         @PathVariable @Min(value = 1, message = ExceptionMessage.BEER_ID_INVALID) Long id) {
+        List<ServerException> exceptions = new LinkedList<>();
         try {
+            Beer beer = beerService.getBeer(id);
+            if (beer == null) {
+                exceptions.add(new ServerException(ExceptionMessage.BEER_NOT_FOUND));
+                return new ResponseEntity<>(exceptions, HttpStatus.NOT_FOUND);
+            }
             beerService.deleteBeer(id);
             return new ResponseEntity<>(id, HttpStatus.OK);
         } catch (Exception e) {
             logger.error(ExceptionMessage.BEER_DELETE_500, e);
-            return new ResponseEntity<>(new ServerException(ExceptionMessage.BEER_DELETE_500),
-                    HttpStatus.INTERNAL_SERVER_ERROR);
+            exceptions.add(new ServerException(ExceptionMessage.BEER_DELETE_500));
+            return new ResponseEntity<>(exceptions, HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
